Move purger to pkg directory
