+++ pkg/api/middlewares.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/weaveworks/common/middleware"
 
-	"github.com/cortexproject/cortex/pkg/storage/purger"
+	"github.com/cortexproject/cortex/pkg/purger"
 	"github.com/cortexproject/cortex/pkg/querier/tripperware/queryrange"
 	"github.com/cortexproject/cortex/pkg/tenant"
 	util_log "github.com/cortexproject/cortex/pkg/util/log"
